+++ classes.py
@@ -21,12 +21,18 @@ class ModelProcessing(str, Enum):
     uvas_de_mesa = "uvas_de_mesa"
     sem_classificacao = "sem_classificacao"
 
+class Token(BaseModel):
+    access_token: str
+    token_type: str
+
+class TokenData(BaseModel):
+    username: Union[str, None] = None
+
 class User(BaseModel):
     username: str
     email: Union[str, None] = None
     full_name: Union[str, None] = None
     disabled: Union[bool, None] = None
-
 
 class UserInDB(User):
     hashed_password: str
